add spec

+++ app/services/web/views/base.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timedelta, timezone
-from typing import List
+from typing import List, Optional, Tuple
 
 import jwt
 from aiohttp import web
@@ -8,7 +8,7 @@ from jwt import ExpiredSignatureError, PyJWTError
 from pydantic import ValidationError, BaseModel
 from tortoise.exceptions import DoesNotExist
 
-from app.api.response import ResponseRefreshAccessToken
+from app.api.response import ResponseRefreshAccessToken, ResponseSuccess
 from app.config import app_config
 from app.consts import Permissions
 from app.models import User, Token
@@ -16,6 +16,8 @@ from app.services.web.exceptions import DataNotFoundException, ApiException, exc
 
 
 class BaseView(web.View):
+    default_success_response = ResponseSuccess()
+
     def __init__(self, request):
         self.uid: int = None
         self.permissions: List[Permissions] = []
@@ -29,34 +31,34 @@ class BaseView(web.View):
             return web.json_response(
                 status=e.status_code,
                 data={
-                    'status': 'fail',
-                    'errors': {'message': exception_codes[e.code], 'code': e.code}
+                    'success': False,
+                    'errors': [{'message': exception_codes[e.code], 'code': e.code}]
                 },
             )
         except ValidationError as e:
             return web.json_response(
                 status=400,
-                data={'status': 'fail', 'errors': {'message': e.errors()}},
+                data={'success': False, 'errors': [{'message': e.errors()}]},
             )
         except web.HTTPClientError as e:
             message = exception_codes.get(e.status_code) or getattr(e, 'text', '')
 
             return web.json_response(
                 status=e.status_code,
-                data={'status': 'fail', 'errors': {'message': message}},
+                data={'success': False, 'errors': [{'message': message}]},
             )
         except DoesNotExist:
             return web.json_response(
                 status=404,
                 data={
-                    'status': 'fail',
-                    'errors': {'message': exception_codes[404], 'code': 404}
+                    'success': False,
+                    'errors': [{'message': exception_codes[404], 'code': 404}]
                 },
             )
         except PyJWTError:
             return web.json_response(
                 status=web.HTTPUnauthorized.status_code,
-                data={'status': 'fail', 'errors': {'message': web.HTTPUnauthorized.text}},
+                data={'success': False, 'errors': [{'message': web.HTTPUnauthorized.text}]},
             )
 
         if isinstance(result, BaseModel):
@@ -91,17 +93,39 @@ class BaseView(web.View):
             raise DataNotFoundException
 
     @staticmethod
-    async def _get_tokens(user: User) -> ResponseRefreshAccessToken:
+    async def _get_tokens(
+            user: User = None,
+            uid: int = None,
+            permissions: List[int] = None,
+    ) -> ResponseRefreshAccessToken:
+
+        if user:
+            uid = user.id
+            permissions = user.permissions
+
+        if not permissions:
+            raise web.HTTPBadRequest
+
         now = datetime.utcnow().replace(tzinfo=timezone.utc)
         access_expire = now + timedelta(seconds=app_config.jwt.access_token_expire)
         refresh_expire = now + timedelta(seconds=app_config.jwt.refresh_token_expire)
 
         access = jwt.encode(
-            {'uid': user.id, 'permissions': user.permissions, 'exp': int(access_expire.timestamp())},
+            {
+                'uid': uid,
+                'permissions': permissions,
+                'exp': int(access_expire.timestamp()),
+                'type': 'access',
+            },
             app_config.jwt.sig,
         )
         refresh = jwt.encode(
-            {'uid': user.id, 'exp': int(refresh_expire.timestamp())},
+            {
+                'uid': uid,
+                'permissions': permissions,
+                'exp': int(refresh_expire.timestamp()),
+                'type': 'refresh',
+            },
             app_config.jwt.sig,
         )
 
@@ -124,15 +148,22 @@ class BaseView(web.View):
         response.set_cookie('refresh_token', refresh, max_age=app_config.jwt.refresh_token_expire)
 
     @staticmethod
-    async def _get_user_by_refresh(refresh: str) -> User:
+    async def _get_user_by_refresh(refresh: str) -> Tuple[Optional[User], List[int]]:
         payload = jwt.decode(refresh, app_config.jwt.sig, algorithms=['HS256'])
+
+        if payload.get('type', '') != 'refresh':
+            raise web.HTTPUnauthorized()
 
         if not await Token.filter(refresh_token=refresh).exists():
             raise web.HTTPUnauthorized()
 
         await Token.filter(refresh_token=refresh).delete()
 
-        return await User.get(id=payload.get('uid'), is_active=True)
+        user = None
+        if payload.get('uid') is not None:
+            user = await User.get(id=payload.get('uid'), is_active=True)
+
+        return user, payload.get('permissions', [])
 
 
 class BaseClientAuthView(BaseView):
@@ -152,13 +183,22 @@ class BaseClientAuthView(BaseView):
 
         try:
             payload = jwt.decode(access, app_config.jwt.sig, algorithms=['HS256'])
+
+            if payload.get('type', '') != 'access':
+                raise web.HTTPUnauthorized()
+
             self.uid = payload.get('uid')
             self.permissions = Permissions.get_permissions(payload.get('permissions', []))
         except ExpiredSignatureError:
-            user = await self._get_user_by_refresh(refresh)
-            self.uid = user.id
-            self.permissions = Permissions.get_permissions(user.permissions)
-            token = await self._get_tokens(user)
+            user, perm = await self._get_user_by_refresh(refresh)
+
+            if user:
+                self.uid = user.id
+                self.permissions = Permissions.get_permissions(user.permissions)
+            else:
+                self.permissions = Permissions.get_permissions(perm)
+
+            token = await self._get_tokens(user, permissions=perm)
             self.request['access_token'] = token.access_token
             self.request['refresh_token'] = token.refresh_token
